CarDto - horsepower && torque added BookedPeriodDto - rename Mapper - fixed

+++ src/main/java/com/telran/ilcarro/repository/entity/PricePerDayEntity.java
@@ -11,5 +11,5 @@ import lombok.*;
 public class PricePerDayEntity {
 
     private String currency;
-    private Double value;
+    private float value;
 }
